chore(web): mass cleanup - removes old contact-point input engine classes, refs
