+++ web/src/engine/osk/src/config/embeddedGestureConfig.ts
@@ -1,7 +1,4 @@
 import GlobeHint from "../globehint.interface.js";
-import PendingGesture from "../input/gestures/pendingGesture.interface.js";
-import { KeyElement } from "../keyElement.js";
-import KeyTip from "../keytip.interface.js";
 import VisualKeyboard from "../visualKeyboard.js";
 
 export default interface EmbeddedGestureConfig {
